requirements_for_ipynb.txt
numpy==1.24.2
faiss-cpu==1.7.4
pandas==2.0.3
matplotlib==3.7.2
scipy==1.11.1
phik==0.12.3
catboost==1.2
optuna==3.3.0
scikit-learn==1.3.0
skimpy==0.0.9
seaborn==0.12.2
tqdm==4.66.1
fastapi==0.101.1
uvicorn==0.23.2
joblib==1.3.2
python-dotenv==1.0.0